fix(jackson): add RefinedShortSerializer for Age serialization
Age implements RefinedShort, not RefinedInt, so the existing
RefinedIntSerializer never matched it. Add RefinedShortSerializer
and register it in RefinedTypeModule.

Also fix test assertions: Age.value() returns short, so
isEqualTo(42) (Integer) always failed — cast to (short).

+++ src/test/java/io/github/dfa1/refinedtype/jackson/RefinedTypeModuleRoundtripTest.java
@@ -112,7 +112,7 @@ class RefinedTypeModuleRoundtripTest {
         Age result = MAPPER.readValue("42", Age.class);
 
         // Then
-        assertThat(result.value()).isEqualTo(42);
+        assertThat(result.value()).isEqualTo((short) 42);
     }
 
     @Test
@@ -143,7 +143,7 @@ class RefinedTypeModuleRoundtripTest {
         assertThat(result.email().value()).isEqualTo("alice@example.com");
         assertThat(result.country().value()).isEqualTo("CH");
         assertThat(result.currency().value()).isEqualTo("CHF");
-        assertThat(result.age().value()).isEqualTo(35);
+        assertThat(result.age().value()).isEqualTo((short) 35);
         assertThat(result.port().value()).isEqualTo(9090);
     }
